+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 * Added a validation that verifies that readme image absolute links do not contain the working branch name.
+* Added a validation to ensure reputation command's default argument is set as an array input.
 
 # 1.5.1
 * Fixed an issue where **validate** command failed to recognized test playbooks for beta integrations as valid tests.